Fix errors on vector lib (#52)
* Fix errors on vector lib

* Fix

+++ libft/srcs/vector/vct_shift_tool.c
@@ -3,34 +3,34 @@
 /*                                                        :::      ::::::::   */
 /*   vct_shift_tool.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cempassi <cempassi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/23 03:12:39 by cempassi          #+#    #+#             */
-/*   Updated: 2020/07/23 03:12:39 by cempassi         ###   ########.fr       */
+/*   Updated: 2021/06/25 21:30:53 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		shift_nright(t_vector *vector, uint64_t start, uint32_t offset)
+void		shift_nright(t_vector *vector, uint64_t start, uint64_t offset)
 {
-	int64_t	vlen;
+	uint64_t	vlen;
 
 	vlen = vct_len(vector);
-	while (vlen >= (int64_t)start)
+	while (vlen >= start)
 	{
 		vector->buffer[vlen + offset] = vector->buffer[vlen];
 		--vlen;
 	}
 }
 
-void		shift_nleft(t_vector *vector, uint64_t start, uint32_t offset)
+void		shift_nleft(t_vector *vector, uint64_t start, uint64_t offset)
 {
 	uint64_t	vlen;
-	int			nb;
+	int64_t			nb;
 
 	vlen = vct_len(vector);
-	nb = vlen - offset;
+	nb = (int64_t)(vlen - offset);
 	while (nb < 0 && offset >= 1)
 		offset--;
 	vlen -= offset;
