+++ libft/srcs/vector/vct_sub.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vct_sub.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cempassi <cempassi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: arsciand <arsciand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/23 03:12:39 by cempassi          #+#    #+#             */
-/*   Updated: 2020/07/23 03:12:39 by cempassi         ###   ########.fr       */
+/*   Updated: 2021/06/25 21:31:40 by arsciand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 char	*vct_sub(t_vector *vector, uint64_t start, uint64_t len)
 {
-	return (ft_strsub(vector->buffer, start, len));
+	return (ft_strsub(vector->buffer, (unsigned int)start, (size_t)len));
 }
